Move `Header` into components folder
