+++ src/components/Layout/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from '../../containers/Header';
+import Header from '../Header';
 import Footer from '../Footer';
 import './Layout.scss';
 import { withRouter } from 'react-router';
